Add HMAC JWT functionality to crypto module
The crypto module has been updated to include functions for creating, signing, and verifying HMAC JWTs. This was accomplished by restructuring the joseAPI and adding new methods in the 'jose.js' file. Also, the required checks have been updated in the 'utils.js' file inside the lib folder of the oauth middleware.